Baddies! And tweaks!
The game is now winnable winnable. i.e. you don't die when you get all the coins. I'm planning on having different levels with different goals, this is the first (money opens doors) and each level will have the final door open when you complete the goal of the level.